+++ com.unity.formats.fbx/Editor/Sources/Recorders/FbxRecorder/FbxRecorderSettings.cs
@@ -91,9 +91,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         /// <returns>Transform associated with binding ID</returns>
         static Transform GetBinding(string id)
         {
-            // return BindingManager.Get(m_BindingId) as GameObject;
-            var method = Type.GetType("UnityEditor.Recorder.BindingManager,Unity.Recorder.Editor").GetMethod("Get", System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.Public);
-            return method.Invoke(null, new object[] { id }) as Transform;
+            return BindingManager.Get(id) as Transform;
         }
 
         /// <summary>
@@ -105,9 +103,7 @@ namespace UnityEditor.Formats.Fbx.Exporter
         /// <param name="obj">Unity Object</param>
         static void SetBinding(string id, UnityEngine.Object obj)
         {
-            // BindingManager.Set(m_BindingId, value);
-            var method = Type.GetType("UnityEditor.Recorder.BindingManager,Unity.Recorder.Editor").GetMethod("Set", System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.Public);
-            method.Invoke(null, new object[] { id, obj });
+            BindingManager.Set(id, obj);
         }
 
         /// <summary>